requirements.txt
cryptography>=41.0.0
openai>=1.0.0
python-dotenv>=1.0.0
gradio>=4.0.0
pydantic>=2.0.0 